requirements.txt
uwsgi==2.0.17.1
pillow==5.3.0
